requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.0.350
openai==0.28.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
faiss-cpu==1.13.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
langchain-text-splitters==0.0.1
voyageai==0.2.4
